+++ .idea/encodings.xml
@@ -3,10 +3,10 @@
   <component name="Encoding">
     <file url="file://$PROJECT_DIR$/azure-openai/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/azure-openai/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/bce-extensions/azure-openai/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/bce-extensions/azure-openai/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/bce-extensions/core/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/bce-extensions/core/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/bce-extensions/langchain4j-azure-openai/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/bce-extensions/langchain4j-azure-openai/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/bce-extensions/langchain4j-core/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/bce-extensions/langchain4j-core/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/bce-extensions/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/bce-extensions/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/examples/helidon-car-booking/src/main/filtered-resources" charset="UTF-8" />
