+++ bce-extensions/langchain4j-azure-openai/src/main/java/io/smallrye/llm/extensions/impl/azureopenai/AzureOpenAiChatModelProducer.java
