+++ bce-extensions/langchain4j-azure-openai/src/main/resources/META-INF/services/jakarta.enterprise.inject.build.compatible.spi.BuildCompatibleExtension
