+++ bce-extensions/langchain4j-core/pom.xml
@@ -7,8 +7,9 @@
         <version>1.0.0-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
-    <artifactId>core</artifactId>
+    <artifactId>langchain4j-core</artifactId>
     <name>SmallRye LLM: Build Compatible Extensions - core</name>
+
 
    <dependencies>
        <dependency>
