+++ bce-extensions/langchain4j-core/src/test/java/io/smallrye/llm/core/ExtensionTest.java
