+++ bce-extensions/pom.xml
@@ -14,8 +14,8 @@
 
 
     <modules>
-        <module>core</module>
-        <module>azure-openai</module>
+        <module>langchain4j-core</module>
+        <module>langchain4j-azure-openai</module>
     </modules>
 
     <dependencies>
