+++ examples/helidon-car-booking/pom.xml
@@ -23,7 +23,7 @@
     <dependencies>
         <dependency>
             <groupId>io.smallrye.llm.bce.extensions</groupId>
-            <artifactId>azure-openai</artifactId>
+            <artifactId>langchain4j-azure-openai</artifactId>
             <version>1.0.0-SNAPSHOT</version>
         </dependency>
         <!-- Adds all features available in MicroProfile -->
