+++ spi/pom.xml
@@ -18,10 +18,5 @@
             <artifactId>jakarta.enterprise.cdi-api</artifactId>
             <scope>provided</scope>
         </dependency>
-        <dependency>
-            <groupId>dev.langchain4j</groupId>
-            <artifactId>langchain4j</artifactId>
-            <scope>compile</scope>
-        </dependency>
     </dependencies>
 </project>
